split up tests for io package
